[586176] Deleting image leaves thumbnail behind.
* Don't save thumbnail twice, wait until it's generated.
* [590004] Fixes foreign key leakage in tests - test_utils.TransactionTestCase truncated django_content_type. Landed a fix in test_utils and added a setting to use it, TEST_UTILS_NO_TRUNCATE
* No more SkipTests in the upload app. Hoping Hudson shows mercy and passes all of them this time.
* Change ImageAttachment docstring to reflect reality.

+++ apps/upload/views.py
@@ -64,7 +64,8 @@ def del_image_async(request, image_id):
             json.dumps({'status': 'error', 'message': message}))
 
     image.file.delete()
-    image.thumbnail.delete()
+    if image.thumbnail:
+        image.thumbnail.delete()
     image.delete()
 
     return HttpResponse(json.dumps({'status': 'success'}))
